Fixed variable BPM, added dual-pack for slower/faster section, added variable gallop and pack speed, cleaned up the automapper algorithm. You can now convert BS map into the automapper.

+++ osuBMParser/TimingPoint.cs
@@ -5,7 +5,7 @@
 
         #region fields
         public int Offset { get; set; }
-        public float MsPerBeat { get; set; }
+        public double MsPerBeat { get; set; }
         public int Meter { get; set; }
         public int SampleType { get; set; }
         public int SampleSet { get; set; }
